Paper over random test failures on 32-bit.
There's a false positive once in hundreds of billions of runs on 64-bit,
once in several runs(!) on 32-bit on a 64-core box.  CI fails to random
crap (failing to fetch from git, setting up image, ...) over 1% of the
time, and the original project doesn't care about necrocomputing.

A proper fix would require some work, and my field of damns is too
barren to give more for 32-bit.